Update lc.c
Changed oriBC to Louise's current version.  This should be identical to the old (original) mpcd with the exception of //make U point out of BC and changing the sign of //Vector between collision point and CM.  (and passing U0 into torqueLCBC).  There is no change yet to torqueLCBC but there will be when Louise has made it working properly.