commenting out exit( 1 ) if cross-time fails
Very occasionally (roughly one particle collision with a non-planar wall in 10^5 or 10^6) creates an invalid crosstime.  Since this is so infrequent, (and if the crosstime is invalid it gets updated to tleft), this will not impact any results - so we don't want long sims exiting if this happens.  Note: this is still an improvement from the previous version that had this broken every time a particle was found inside a non-planar wall.  Eventually we should look into why this happens.